+++ src/main/java/com/github/antonfedoruk/boardgamesgooglesheettgbot/command/HelpCommand.java
@@ -1,6 +1,8 @@
 package com.github.antonfedoruk.boardgamesgooglesheettgbot.command;
 
 import com.github.antonfedoruk.boardgamesgooglesheettgbot.service.SendBotMessageService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
 import static com.github.antonfedoruk.boardgamesgooglesheettgbot.command.CommandName.*;
@@ -9,6 +11,7 @@ import static com.github.antonfedoruk.boardgamesgooglesheettgbot.command.Command
  * Help {@link Command}.
  */
 public class HelpCommand implements Command {
+    private static final Logger log = LoggerFactory.getLogger(HelpCommand.class);
     private final SendBotMessageService sendBotMessageService;
 
     public static final String HELP_MESSAGE = String.format("<b>Дотупні команди</b>\n"
@@ -32,5 +35,10 @@ public class HelpCommand implements Command {
     @Override
     public void execute(Update update) {
         sendBotMessageService.sendMessage(CommandUtils.getChatId(update), HELP_MESSAGE);
+    }
+
+    @Override
+    public Logger getLogger() {
+        return log;
     }
 }