+++ src/main/java/com/github/antonfedoruk/boardgamesgooglesheettgbot/command/StartCommand.java
@@ -3,6 +3,8 @@ package com.github.antonfedoruk.boardgamesgooglesheettgbot.command;
 import com.github.antonfedoruk.boardgamesgooglesheettgbot.repository.entity.TelegramUser;
 import com.github.antonfedoruk.boardgamesgooglesheettgbot.service.SendBotMessageService;
 import com.github.antonfedoruk.boardgamesgooglesheettgbot.service.TelegramUserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
 import static com.github.antonfedoruk.boardgamesgooglesheettgbot.command.CommandName.HELP;
@@ -13,6 +15,7 @@ import static com.github.antonfedoruk.boardgamesgooglesheettgbot.command.Command
  * Start {@link Command}.
  */
 public class StartCommand implements Command {
+    private static final Logger log = LoggerFactory.getLogger(StartCommand.class);
     private final SendBotMessageService sendBotMessageService;
     private final TelegramUserService telegramUserService;
 
@@ -40,5 +43,10 @@ public class StartCommand implements Command {
                     telegramUserService.save(telegramUser);
                 });
         sendBotMessageService.sendMessage(getChatId(update), START_MESSAGE);
+    }
+
+    @Override
+    public Logger getLogger() {
+        return log;
     }
 }